+++ packages/dom/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @page-ai/dom
 
+## 0.9.0
+
+### Minor Changes
+
+- Update to had actual click handlers and interactivity
+
+## 0.8.0
+
+### Minor Changes
+
+- Fix react peerDeps
+
+## 0.7.0
+
+### Minor Changes
+
+- Customer serializer
+
+## 0.6.0
+
+### Minor Changes
+
+- Patch more bugs
+
+## 0.5.0
+
+### Minor Changes
+
+- win.location patch #2
+
+## 0.4.0
+
+### Minor Changes
+
+- Fix win.location bug #1
+
+## 0.3.0
+
+### Minor Changes
+
+- Update support for Next.js
+
+## 0.2.0
+
+### Minor Changes
+
+- Attempt to get types to generate
+
 ## 0.1.0
 
 ### Minor Changes