+++ packages/react/CHANGELOG.md
@@ -1,5 +1,107 @@
 # @page-ai/react
 
+## 0.12.0
+
+### Minor Changes
+
+- Update to had actual click handlers and interactivity
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.9.0
+  - @page-ai/dom@0.9.0
+
+## 0.11.0
+
+### Minor Changes
+
+- Fix react peerDeps
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.8.0
+  - @page-ai/dom@0.8.0
+
+## 0.10.0
+
+### Minor Changes
+
+- Customer serializer
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.7.0
+  - @page-ai/dom@0.7.0
+
+## 0.9.0
+
+### Minor Changes
+
+- Patch more bugs
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.6.0
+  - @page-ai/dom@0.6.0
+
+## 0.8.0
+
+### Minor Changes
+
+- win.location patch #2
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.5.0
+  - @page-ai/dom@0.5.0
+
+## 0.7.0
+
+### Minor Changes
+
+- Fix win.location bug #1
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.4.0
+  - @page-ai/dom@0.4.0
+
+## 0.6.0
+
+### Minor Changes
+
+- Update support for Next.js
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.3.0
+  - @page-ai/dom@0.3.0
+
+## 0.5.0
+
+### Minor Changes
+
+- Attempt to get types to generate
+
+### Patch Changes
+
+- Updated dependencies
+  - @page-ai/core@0.2.0
+  - @page-ai/dom@0.2.0
+
+## 0.4.0
+
+### Minor Changes
+
+- Working on fixing versions
+
 ## 0.3.0
 
 ### Minor Changes